pyproject.toml
[project]
name = "goit-pythonweb-hw-10"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmytro Popovskyi",email = "popovskiy96@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "phonenumbers (>=9.0.18,<10.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "faker (>=38.0.0,<39.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "bcrypt (<4.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "fastapi-mail (>=1.5.8,<2.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.44.1,<2.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "alembic (>=1.17.2,<2.0.0)"
]
